// client_solid/src/generated/sdui.ts
export type AnalyticsEvent = {
  analyticsId: string,
  eventType: AnalyticsEventType,
  timestampMs: number,
  context?: { [key: string]: string },
};

// ── AnalyticsEventType.ts ───────────────────────────────────────
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AnalyticsEventType = "impression" | "click";